Refactor services structure and update import paths
