+++ src/components/detail/RecipeCard.tsx
@@ -5,7 +5,7 @@ import {
 } from '@mui/material';
 import type { Recipe } from '../../types/index';
 import FactorioIcon from '../common/FactorioIcon';
-import { DataService } from '../../services/DataService';
+import { DataService } from '../../services/core/DataService';
 import useGameStore from '../../store/gameStore';
 import CraftingButtons from './CraftingButtons';
