+++ src/components/detail/RecipeFlowDisplay.tsx
@@ -6,7 +6,7 @@ import { Add as AddIcon, ArrowForward as ArrowIcon } from '@mui/icons-material';
 import type { Recipe } from '../../types/index';
 import FactorioIcon from '../common/FactorioIcon';
 import useGameStore from '../../store/gameStore';
-import { DataService } from '../../services/DataService';
+import { DataService } from '../../services/core/DataService';
 import TimeIcon from '../../assets/Time.png';
 
 interface RecipeFlowDisplayProps {
