+++ src/components/technology/TechGridCard.tsx
@@ -16,8 +16,8 @@ import {
   NewReleases as TriggerIcon
 } from '@mui/icons-material';
 import FactorioIcon from '../common/FactorioIcon';
-import { TechnologyService } from '../../services/TechnologyService';
-import { DataService } from '../../services/DataService';
+import { TechnologyService } from '../../services/core/TechnologyService';
+import { DataService } from '../../services/core/DataService';
 import type { Technology, TechStatus } from '../../types/technology';
 
 // Factorio Design System
