+++ src/hooks/useServices.ts
@@ -4,7 +4,7 @@
  */
 
 import { useMemo } from 'react';
-import { ServiceLocator, SERVICE_NAMES } from '../services/ServiceLocator';
+import { ServiceLocator, SERVICE_NAMES } from '../services/utils/ServiceLocator';
 import { DataService } from '../services/core/DataService';
 import { RecipeService } from '../services/core/RecipeService';
 import { TechnologyService } from '../services/core/TechnologyService';
@@ -12,7 +12,7 @@ import { FuelService } from '../services/FuelService';
 import { PowerService } from '../services/game/PowerService';
 import { StorageService } from '../services/game/StorageService';
 import { UserProgressService } from '../services/state/UserProgressService';
-import type { GameStateProvider } from '../services/interfaces';
+import type { GameStateProvider } from '../services/interfaces/index';
 
 /**
  * 获取 DataService 实例
