+++ src/utils/manualCraftingValidator.ts
@@ -1,7 +1,7 @@
 // 手动采集验证工具类
 // 基于 Factorio Wiki 官方规则实现
 
-import { ServiceLocator, SERVICE_NAMES } from '../services/ServiceLocator';
+import { ServiceLocator, SERVICE_NAMES } from '../services/utils/ServiceLocator';
 import type { DataService } from '../services/core/DataService';
 import { RecipeService } from '../services/core/RecipeService';
 import type { Recipe } from '../types/index';
